requirements.txt
asgiref==3.5.0
certifi==2021.5.30
charset-normalizer==2.0.6
Django==4.0.2
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
idna==3.2
importlib-metadata==4.10.1
Markdown==3.3.6
mysqlclient==2.1.0
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.7
zipp==3.7.0

PyJWT~=2.3.0